requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
python-docx>=0.8.11

# PDF handling (report content extraction)
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Web scraping dependencies
playwright>=1.48.0
playwright-stealth>=1.0.6
markdownify>=0.11.6
pdfkit>=1.0.0
fake-useragent>=1.5.1

# LLM API dependencies
openai>=1.0.0

# GUI dependencies
customtkinter>=5.2.0

# Build/packaging dependencies
pyinstaller>=6.0.0
